requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.40.0
openai>=1.30.0
python-dotenv==1.0.1
requests==2.31.0
